+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+
+# 0.6.1 (2018-03-13)
+
+- DashReactTable columns now filterable using a comma-separated list of values
+- DashReactSelect now updates `this.props.value` so state changes are propagated to Dash
+
+
 # 0.6.0 (2018-03-12)
 
 - Added new DashReactSelect component